+++ bro_config_manager/lib/bro_config_manager.dart
@@ -6,8 +6,7 @@ library;
 
 export 'src/models/config_var/config_var.dart';
 export 'src/models/config_var/config_var_list.dart';
-export 'src/models/config_var/config_var_list_with_default.dart';
-export 'src/models/config_var/config_var_with_default.dart';
+export 'src/models/config_var/simple_config_var.dart';
 export 'src/models/errors/config_file_error.dart';
 export 'src/models/errors/config_var_not_found_error.dart';
 export 'src/models/init_config_manager_model.dart';