+++ bro_global_manager/lib/src/models/global_manager_not_created_error.dart
@@ -1,0 +1,4 @@
+class GlobalManagerNotCreatedError extends Error {
+  @override
+  String toString() => "The global manager was not created.";
+}