+++ bro_logger_manager/lib/bro_logger_manager.dart
@@ -1,0 +1,8 @@
+// SPDX-FileCopyrightText: 2025 Benoit Rolandeau <borlnov.obsessio@gmail.com>
+//
+// SPDX-License-Identifier: MIT
+
+library;
+
+export 'src/mixins/mixin_logger_configs.dart';
+export 'src/services/logger_manager.dart';