+++ bro_config_manager/CHANGELOG.md
@@ -7,7 +7,11 @@ SPDX-License-Identifier: MIT
 <!-- We disable the rule for having the top level at first line, because we prefer a second level
      header for version -->
 <!-- markdownlint-disable MD041 -->
+## 1.1.0
+
+- Refactor the config variable classes to add a value converter in it.
+- Update logger dependencies
+
 ## 1.0.0
-<!-- markdownlint-enable MD041 -->
 
 - Initial release.